Don't save library version to NSUserDefaults (#212)
This seems to be a holdover from some old code, and it's not being used anywhere in the component. Plus it's better not to be saving anything in `[NSUserDefaults standardUserDefaults]` as that is the developer/app's domain.

Also renamed the messaging code while I was in there to be something meaningful ("000" → "PrintLibraryVersion").